+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* Added reload button to transaction history page @faboweb
+
+### Changed
+
+* Showing uniq transactions by their hash in history @faboweb
+
 ### Fixed
 
 * fixed block explorer always showing "syncing" @faboweb